pyproject.toml
[project]
name = "genesis-humanoid-rl"
version = "0.1.0"
description = "Humanoid robot reinforcement learning with Genesis physics engine and Stable-Baselines3"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "torch>=2.0.0",
    "genesis-world",
    "stable-baselines3[extra]>=2.0.0",
    "numpy>=1.21.0",
    "gymnasium>=0.29.0",
    "matplotlib>=3.5.0",
    "tensorboard>=2.8.0",
    "trimesh>=4.6.12",
    "wandb>=0.15.0",
    "pytest>=8.4.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "psutil>=5.9.0",
    "python-multipart>=0.0.6",
    "pytest-asyncio>=1.0.0",
    "imageio[ffmpeg]>=2.37.0",
    "gputil>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.991",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
strict = true
